backend/NotificationsService/requirements.txt
flask
flask-cors
requests
datetime
mysql-connector-python==9.0.0